cloud_run/requirements.txt
Flask==3.0.3
google-cloud-pubsub==2.21.5